refactor(viewer): move colorArrayToFn to colormap.js
colorArrayToFn is a color transformation utility, not a DOM concern.
Move it from color_legend.js to util/colormap.js alongside the other
color ramp functions (infernoColor, interpolateRamp).
